Labels and Trendline
Added label for the y-axis and a title for the plot. Also figured out how to incorporate a trendline into the plot so I set that up and typed comments throughout the code to 1) keep track of what the code is doing and 2) ensure my understanding with what it is doing and how it works.