Plot moved and Spacing
Moved the velocity vs time test plot to accelerated motion section because the graph is more realistic as the slope would be acceleration. Also removed the trendline-related code for this plot as it doesn't work with how it is currently set up. Lastly, I added empty print statements at the end of several of the codes to allow easier readability throughout the lab