Sawtooth Plot
Commented out the "df.info()" so that the lab doesn't seem as cluttered. Added a mock csv file for the sawtooth graph to replace the first position vs time graph to give a more realistic example of what plot should be there. Edited the "trendline" to be a line that connects the dots in a way that simulates the "sawtooth" shape of the plot that the student should be aiming for.