Raw Data 2 and 3
Was going to have separate cells to minimize the tedious nature with copying and pasting and changing each of the numbers, but I'm not sure it would make much difference. Fixed spot where it still said "cylinder number". Added raw data 2 for cart 2. I'm not sure if we want them to find the measure acceleration themselves or let the code do it for them like it currently is. Also added Raw data 3 and changed data to 0 since the code appears to work.