Raw data start
Started code for raw data tables in experiment 1. Copied most of it from lab 4, but I realized the units for the uncertainty of velocity are wrong so I need to fix that. Additionally I am getting a different number for the uncertainty of velocity calculation so I need to see if it is because of my own error from the lab or in my code.