Lab 5 Plot and Lines
Started code for Initial Potential Energy vs. Final Kinetic Energy plot (need to do error bars). Also changed lines separating parts of the lab to double lines to be able to keep which code is for which experiment it is for more organized. Deleted code cell that I think was pasted from lab 4 by accident.