Lab 4 several code changes
Added print(df) statement instead of needing code for table. Added statement to add uncertainty of am to dataframe to look more like the original table. The dataframe might have an extra column in it depending on the capstone software, be aware of this for trouble shooting. Going to copy code and delete the now un-needed code.